Added Flippers/Blue Key Check to W3 Merciless Logic
The Jellyfish Room in W3 now checks if the region contains specifically the Blue Key, or if the player has Flippers. Just like with the Spark Room in N6, it's impossible to escape once the player exits the water. Therefore, any Key Color besides the Blue will logically require Flippers to obtain.