W3 Jellyfish Room Logic Update
Tested this again. It is possible to escape using the same water treading trick as in N6. The block will be down, thus blocking the door again. However, you can get out of the block by simply moving right. As such, this logic has now been updated so that it checks for any key color besides the Green. The Green Key will still logically require Flippers, but that should be the only exception.